Controller

+++ src/main/java/com/SchoolEmployees/Services/EmployeeServices.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.SchoolEmployees.Domain.Employee;
+import com.SchoolEmployees.Exceptions.NotFoundException;
 import com.SchoolEmployees.Repo.EmployeeRepo;
 
 @Service
@@ -16,14 +17,13 @@ public class EmployeeServices {
 	private EmployeeRepo repo;
 
 	public EmployeeServices(EmployeeRepo repo) {
-		super();
 		this.repo = repo;
 	
 	}
 	
 	//get an employee
 	public Employee getEmployee(long id) {
-		return this.repo.findById(id).get();
+		return this.repo.findById(id).orElseThrow(() -> new NotFoundException("employee with id " + id + " not found" ));
 	}
 	
 	public Employee createEmployee(Employee employee) {
@@ -35,7 +35,7 @@ public class EmployeeServices {
 	}
 	//Adding a new employee
 	public Employee udpateEmployee(Employee newEmployee, long id) {
-		Employee found = repo.findById(id).get();
+		Employee found = repo.findById(id).orElseThrow();
 		
 		found.setFirstName(newEmployee.getFirstName());
 		found.setLastName(newEmployee.getLastName());
@@ -47,6 +47,11 @@ public class EmployeeServices {
 	}
 	
 	public void deleteEmployee(long id) {
+		var isEmployeeNotPresent = repo.findById(id).isEmpty();
+		if (isEmployeeNotPresent) {
+			throw new NotFoundException("employee with id " + id + " not found" );
+		
+		}	
 		repo.deleteById(id);
 		
 	}
